Created several functional component files and connected them to redux
